refactor: config 클래스 패키지로 분류
