+++ backend/src/main/java/com/woowacourse/zzimkkong/config/SlackConfig.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzimkkong;
+package com.woowacourse.zzimkkong.config;
 
 import com.woowacourse.zzimkkong.domain.SlackUrl;
 import org.springframework.beans.factory.annotation.Value;
